10K_Text_Analysis/requirements.txt
requests
pandas
python-dotenv
tqdm